Add print keyword
We are intentionally not adding an op_print opcode.
The print atom is intended for debugging in a development environment.
print passes through the value of its reduced arguments, so it can be inserted anywhere in the tree.
'brun -p' and 'run -p' will act as if print is applied to all reductions.
The tests for the print keyword work by re-running all tests with the '-p' flag,
and checking that the output is identical.